+++ engine/src/frog/gx3d/location.hpp
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "frog/gl/include.hpp"
+#include "frog/gx3d/include.hpp"
 
 
 namespace frog::gl::location
