+++ engine/src/frog/os/window.cpp
@@ -1,7 +1,7 @@
 #ifndef NOT_FROG_BUILD_3D
 
 
-#include "frog/gl/include.hpp"
+#include "frog/gx3d/include.hpp"
 
 #include "window.hpp"
