+++ game/CMakeLists.txt
@@ -20,7 +20,7 @@ set(SRC_FILES
 include_directories(${FROG_INCLUDE_DIRS})
 
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED on)
 
 if(NOT MSVC) # settings for the normal compilers
